refactor: 로그 아웃 테스트 코드 수정

+++ src/main/java/balancetalk/module/member/application/MemberService.java
@@ -116,6 +116,9 @@ public class MemberService {
     @Transactional
     public void logout(){
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication == null) {
+            throw new BalanceTalkException(ErrorCode.AUTHENTICATION_REQUIRED);
+        }
         String username = authentication.getName();
         redisService.deleteValues(username);
     }
